+++ pp/cell.py
@@ -107,13 +107,17 @@ def cell(
         #         f"Args {keys} have repeated first letters {first_letters}"
         #     )
 
-        # if "args" not in sig.parameters and "kwargs" not in sig.parameters:
-        #     for key in kwargs.keys():
-        #         if key not in sig.parameters.keys():
-        #             raise TypeError(
-        #                 f"{component_type}() got invalid argument `{key}`\n"
-        #                 f"valid arguments are {list(sig.parameters.keys())}"
-        #             )
+        if (
+            "args" not in sig.parameters
+            and "kwargs" not in sig.parameters
+            and "waveguide_settings" not in sig.parameters
+        ):
+            for key in kwargs.keys():
+                if key not in sig.parameters.keys():
+                    raise TypeError(
+                        f"{component_type}() got invalid argument `{key}`\n"
+                        f"valid arguments are {list(sig.parameters.keys())}"
+                    )
 
         # print(CACHE.keys())
         # print(name)