+++ pp/tests/test_netlists.py
@@ -85,7 +85,7 @@ if __name__ == "__main__":
     # pp.show(c2)
 
     test_netlists(
-        component_type="ring_single",
+        component_type="mzi",
         full_settings=True,
         data_regression=None,
         check=False,